+++ app/environments/brassbirmingham/brassbirmingham/envs/classes/roads/canal.py
@@ -1,0 +1,7 @@
+from .road import Road
+
+
+class Canal(Road):
+    def __init__(self, owner):
+        super(Canal, self).__init__(owner, "canal")
+        self.cost = 3
