+++ app/environments/brassbirmingham/brassbirmingham/envs/classes/roads/railroad.py
@@ -1,0 +1,7 @@
+from .road import Road
+
+
+class Railroad(Road):
+    def __init__(self, owner):
+        super(Railroad, self).__init__(owner, "railroad")
+        # TODO cost
