+++ app/environments/brassbirmingham/brassbirmingham/envs/consts.py
@@ -1,0 +1,420 @@
+from .classes.buildings.industry_building import IndustryBuilding
+from .classes.buildings.market_building import MarketBuilding
+from .classes.town import Town
+from .classes.build_location import BuildLocation
+from .classes.trade_post import TradePost
+from .classes.road_location import RoadLocation
+from .classes.cards.location_card import LocationCard
+from .classes.cards.industry_card import IndustryCard
+
+STARTING_ROADS = 14
+STARTING_MONEY = 17
+# TOTAL_COAL = 30
+# TOTAL_IRON = 18
+# TOTAL_BEER = 15
+# amt of tokens doesn't matter actually
+
+TOWNS = [
+    Town(
+        "blue",
+        "Leek",
+        [BuildLocation(["cotton", "goods"]), BuildLocation(["cotton", "coal"])],
+    ),
+    Town(
+        "blue",
+        "Stoke-On-Trent",
+        [
+            BuildLocation(["cotton", "goods"]),
+            BuildLocation(["pottery", "oil"]),
+            BuildLocation(["goods"]),
+        ],
+    ),
+    Town(
+        "blue",
+        "Stone",
+        [BuildLocation(["cotton", "beer"]), BuildLocation(["goods", "coal"])],
+    ),
+    Town(
+        "blue",
+        "Uttoxeter",
+        [BuildLocation(["goods", "beer"]), BuildLocation(["cotton", "beer"])],
+    ),
+    Town(
+        "green",
+        "Belper",
+        [
+            BuildLocation(["cotton", "goods"]),
+            BuildLocation(["coal"]),
+            BuildLocation(["pottery"]),
+        ],
+    ),
+    Town(
+        "green",
+        "Derby",
+        [
+            BuildLocation(["cotton", "beer"]),
+            BuildLocation(["cotton", "goods"]),
+            BuildLocation(["oil"]),
+        ],
+    ),
+    Town(
+        "red",
+        "Stafford",
+        [BuildLocation(["goods", "beer"]), BuildLocation(["pottery"])],
+    ),
+    Town(
+        "red",
+        "Burton-Upon-Trent",
+        [BuildLocation(["goods", "coal"]), BuildLocation(["beer"])],
+    ),
+    Town("beer1", "", [BuildLocation(["beer"])]),
+    Town("red", "Cannock", [BuildLocation(["goods", "coal"]), BuildLocation(["coal"])]),
+    Town(
+        "red",
+        "Tamworth",
+        [BuildLocation(["cotton", "coal"]), BuildLocation(["cotton", "coal"])],
+    ),
+    Town(
+        "red",
+        "Walsall",
+        [BuildLocation(["oil", "goods"]), BuildLocation(["goods", "beer"])],
+    ),
+    Town(
+        "yellow",
+        "Coalbrookdale",
+        [
+            BuildLocation(["oil", "beer"]),
+            BuildLocation(["oil"]),
+            BuildLocation(["coal"]),
+        ],
+    ),
+    Town(
+        "yellow",
+        "Wolverhampton",
+        [BuildLocation(["goods"]), BuildLocation(["goods", "coal"])],
+    ),
+    Town("yellow", "Dudley", [BuildLocation(["coal"]), BuildLocation(["oil"])]),
+    Town(
+        "yellow",
+        "Kidderminster",
+        [BuildLocation(["cotton", "coal"]), BuildLocation(["cotton"])],
+    ),
+    Town("beer2", "", [BuildLocation(["beer"])]),
+    Town("yellow", "Worcester", [BuildLocation(["cotton"]), BuildLocation(["cotton"])]),
+    Town(
+        "purple",
+        "Birmingham",
+        [
+            BuildLocation(["cotton", "goods"]),
+            BuildLocation(["goods"]),
+            BuildLocation(["oil"]),
+            BuildLocation(["goods"]),
+        ],
+    ),
+    Town(
+        "purple",
+        "Nuneaton",
+        [BuildLocation(["goods", "beer"]), BuildLocation(["cotton", "coal"])],
+    ),
+    Town(
+        "purple",
+        "Coventry",
+        [
+            BuildLocation(["pottery"]),
+            BuildLocation(["goods", "coal"]),
+            BuildLocation(["oil", "goods"]),
+        ],
+    ),
+    Town(
+        "purple", "Redditch", [BuildLocation(["goods", "coal"]), BuildLocation(["oil"])]
+    ),
+]
+
+TRADEPOSTS = [
+    TradePost("Warrington", 2, 5, 0, 0, 2, False),
+    TradePost("Nottingham", 2, 0, 3, 0, 2, False),
+    TradePost("Shrewbury", 1, 0, 4, 0, 2, False),
+    TradePost("Oxford", 2, 0, 0, 2, 2, False),
+    TradePost("Gloucester", 2, 0, 0, 2, 2, True),
+]
+
+ROAD_LOCATIONS = [
+    RoadLocation(["Warrington", "Stoke-On-Trent"]),
+    RoadLocation(["Stoke-On-Trent", "Leek"]),
+    RoadLocation(["Leek", "Belper"], False),
+    RoadLocation(["Belper", "Derby"]),
+    RoadLocation(["Derby", "Nottingham"]),
+    RoadLocation(["Derby", "Uttoxeter"], False),
+    RoadLocation(["Derby", "Burton-Upon-Trent"]),
+    RoadLocation(["Stoke-On-Trent", "Stone"]),
+    RoadLocation(["Stone", "Uttoxeter"], False),
+    RoadLocation(["Stone", "Stafford"]),
+    RoadLocation(["Stone", "Burton-Upon-Trent"]),
+    RoadLocation(["Stafford", "Cannock"]),
+    RoadLocation(["Cannock", "Burton-Upon-Trent"], False),
+    RoadLocation(["Tamworth", "Burton-Upon-Trent"]),
+    RoadLocation(["Walsall", "Burton-Upon-Trent"], canBuildRailroad=False),
+    RoadLocation(["beer1", "Cannock"]),
+    RoadLocation(["Wolverhampton", "Cannock"]),
+    RoadLocation(["Walsall", "Cannock"]),
+    RoadLocation(["Wolverhampton", "Coalbrookdale"]),
+    RoadLocation(["Shrewbury", "Coalbrookdale"]),
+    RoadLocation(["Kidderminster", "Coalbrookdale"]),
+    RoadLocation(["Kidderminster", "Dudley"]),
+    RoadLocation(["Wolverhampton", "Walsall"]),
+    RoadLocation(["Wolverhampton", "Dudley"]),
+    RoadLocation(["Tamworth", "Walsall"], False),
+    RoadLocation(["Nuneaton", "Walsall"]),
+    RoadLocation(["Nuneaton", "Coventry"]),
+    RoadLocation(["Cannock", "Walsall"]),
+    RoadLocation(["Birmingham", "Walsall"]),
+    RoadLocation(["Birmingham", "Tamworth"]),
+    RoadLocation(["Birmingham", "Nuneaton"], False),
+    RoadLocation(["Birmingham", "Coventry"]),
+    RoadLocation(["Birmingham", "Oxford"]),
+    RoadLocation(["Birmingham", "Redditch"], False),
+    RoadLocation(["Birmingham", "Worcester"]),
+    RoadLocation(["Birmingham", "Dudley"]),
+    RoadLocation(["Redditch", "Oxford"]),
+    RoadLocation(["Redditch", "Gloucester"]),
+    RoadLocation(["Worcester", "Gloucester"]),
+    RoadLocation(["Worcester", "beer2", "Kidderminster"]),
+]
+
+BUILDINGS = [
+    MarketBuilding("goods", 1, 8, 1, 0, 1, 3, 5, 2, onlyPhaseOne=True),
+    MarketBuilding("goods", 2, 10, 0, 1, 1, 5, 0, 1),
+    MarketBuilding("goods", 2, 10, 0, 1, 1, 5, 0, 1),
+    MarketBuilding("goods", 3, 12, 2, 0, 0, 4, 4, 0),
+    MarketBuilding("goods", 4, 8, 0, 1, 1, 3, 6, 1),
+    MarketBuilding("goods", 5, 16, 1, 0, 2, 8, 2, 2),
+    MarketBuilding("goods", 5, 16, 1, 0, 2, 8, 2, 2),
+    MarketBuilding("goods", 6, 20, 0, 0, 1, 7, 6, 1),
+    MarketBuilding("goods", 7, 16, 1, 1, 0, 9, 4, 0),
+    MarketBuilding("goods", 8, 20, 0, 2, 1, 11, 1, 1),
+    MarketBuilding("cotton", 1, 12, 0, 0, 1, 5, 5, 1, onlyPhaseOne=True),
+    MarketBuilding("cotton", 1, 12, 0, 0, 1, 5, 5, 1, onlyPhaseOne=True),
+    MarketBuilding("cotton", 1, 12, 0, 0, 1, 5, 5, 1, onlyPhaseOne=True),
+    MarketBuilding("cotton", 2, 14, 1, 0, 1, 5, 4, 2),
+    MarketBuilding("cotton", 2, 14, 1, 0, 1, 5, 4, 2),
+    MarketBuilding("cotton", 3, 16, 1, 1, 1, 9, 3, 1),
+    MarketBuilding("cotton", 3, 16, 1, 1, 1, 9, 3, 1),
+    MarketBuilding("cotton", 3, 16, 1, 1, 1, 9, 3, 1),
+    MarketBuilding("cotton", 4, 18, 1, 1, 1, 12, 2, 1),
+    MarketBuilding("cotton", 4, 18, 1, 1, 1, 12, 2, 1),
+    MarketBuilding("cotton", 4, 18, 1, 1, 1, 12, 2, 1),
+    MarketBuilding("pottery", 1, 17, 0, 1, 1, 10, 5, 1, canBeDeveloped=False),
+    MarketBuilding("pottery", 2, 0, 1, 0, 1, 1, 1, 1),
+    MarketBuilding("pottery", 3, 22, 2, 0, 2, 11, 5, 1, canBeDeveloped=False),
+    MarketBuilding("pottery", 4, 0, 1, 0, 1, 1, 1, 1),
+    MarketBuilding("pottery", 5, 24, 2, 0, 2, 20, 5, 1, onlyPhaseTwo=True),
+    IndustryBuilding("oil", 1, 4, 5, 1, 0, 3, 3, 1, onlyPhaseOne=True),
+    IndustryBuilding("oil", 2, 4, 7, 1, 0, 5, 3, 1),
+    IndustryBuilding("oil", 3, 5, 9, 1, 0, 7, 2, 1),
+    IndustryBuilding("oil", 4, 6, 12, 1, 0, 9, 1, 1),
+    IndustryBuilding("beer", 1, 1, 5, 0, 1, 4, 4, 2),
+    IndustryBuilding("beer", 1, 1, 5, 0, 1, 4, 4, 2),
+    IndustryBuilding(
+        "beer", 2, 1, 7, 0, 1, 5, 5, 2
+    ),  # add logic somewhere to add +1 beer to tier ^2 in second phase
+    IndustryBuilding("beer", 2, 1, 7, 0, 1, 5, 5, 2),
+    IndustryBuilding("beer", 3, 1, 9, 0, 1, 7, 5, 2),
+    IndustryBuilding("beer", 3, 1, 9, 0, 1, 7, 5, 2),
+    IndustryBuilding("beer", 4, 1, 9, 0, 1, 10, 5, 2),
+    IndustryBuilding("coal", 1, 2, 5, 0, 0, 1, 4, 2, onlyPhaseOne=True),
+    IndustryBuilding("coal", 2, 3, 7, 0, 0, 2, 7, 1),
+    IndustryBuilding("coal", 2, 3, 7, 0, 0, 2, 7, 1),
+    IndustryBuilding("coal", 3, 4, 8, 0, 1, 3, 6, 1),
+    IndustryBuilding("coal", 3, 4, 8, 0, 1, 3, 6, 1),
+    IndustryBuilding("coal", 4, 5, 10, 0, 1, 4, 5, 1),
+    IndustryBuilding("coal", 4, 5, 10, 0, 1, 4, 5, 1),
+]
+
+"""
+Starting deck
+
+Key is (str) amount of players playing
+"""
+STARTING_CARDS = {
+    "2": [
+        LocationCard("Stafford"),
+        LocationCard("Stafford"),
+        LocationCard("Burton-Upon-Trent"),
+        LocationCard("Burton-Upon-Trent"),
+        LocationCard("Cannock"),
+        LocationCard("Cannock"),
+        LocationCard("Tamworth"),
+        LocationCard("Walsall"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Dudley"),
+        LocationCard("Dudley"),
+        LocationCard("Kidderminster"),
+        LocationCard("Kidderminster"),
+        LocationCard("Wolverhampton"),
+        LocationCard("Wolverhampton"),
+        LocationCard("Worcester"),
+        LocationCard("Worcester"),
+        LocationCard("Birmingham"),
+        LocationCard("Birmingham"),
+        LocationCard("Birmingham"),
+        LocationCard("Coventry"),
+        LocationCard("Coventry"),
+        LocationCard("Coventry"),
+        LocationCard("Nuneaton"),
+        LocationCard("Redditch"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Coal Mine"),
+        IndustryCard("Coal Mine"),
+        IndustryCard("Pottery"),
+        IndustryCard("Pottery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+    ],
+    "3": [
+        LocationCard("Leek"),
+        LocationCard("Leek"),
+        LocationCard("Stoke-On-Trent"),
+        LocationCard("Stoke-On-Trent"),
+        LocationCard("Stoke-On-Trent"),
+        LocationCard("Stone"),
+        LocationCard("Stone"),
+        LocationCard("Uttoxeter"),
+        LocationCard("Stafford"),
+        LocationCard("Stafford"),
+        LocationCard("Burton-Upon-Trent"),
+        LocationCard("Burton-Upon-Trent"),
+        LocationCard("Cannock"),
+        LocationCard("Cannock"),
+        LocationCard("Tamworth"),
+        LocationCard("Walsall"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Dudley"),
+        LocationCard("Dudley"),
+        LocationCard("Kidderminster"),
+        LocationCard("Kidderminster"),
+        LocationCard("Wolverhampton"),
+        LocationCard("Wolverhampton"),
+        LocationCard("Worcester"),
+        LocationCard("Worcester"),
+        LocationCard("Birmingham"),
+        LocationCard("Birmingham"),
+        LocationCard("Birmingham"),
+        LocationCard("Coventry"),
+        LocationCard("Coventry"),
+        LocationCard("Coventry"),
+        LocationCard("Nuneaton"),
+        LocationCard("Redditch"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Coal Mine"),
+        IndustryCard("Coal Mine"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Pottery"),
+        IndustryCard("Pottery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+    ],
+    "4": [
+        LocationCard("Belper"),
+        LocationCard("Belper"),
+        LocationCard("Derby"),
+        LocationCard("Derby"),
+        LocationCard("Derby"),
+        LocationCard("Leek"),
+        LocationCard("Leek"),
+        LocationCard("Stoke-On-Trent"),
+        LocationCard("Stoke-On-Trent"),
+        LocationCard("Stoke-On-Trent"),
+        LocationCard("Stone"),
+        LocationCard("Stone"),
+        LocationCard("Uttoxeter"),
+        LocationCard("Uttoxeter"),
+        LocationCard("Stafford"),
+        LocationCard("Stafford"),
+        LocationCard("Burton-Upon-Trent"),
+        LocationCard("Burton-Upon-Trent"),
+        LocationCard("Cannock"),
+        LocationCard("Cannock"),
+        LocationCard("Tamworth"),
+        LocationCard("Walsall"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Coalbrookdale"),
+        LocationCard("Dudley"),
+        LocationCard("Dudley"),
+        LocationCard("Kidderminster"),
+        LocationCard("Kidderminster"),
+        LocationCard("Wolverhampton"),
+        LocationCard("Wolverhampton"),
+        LocationCard("Worcester"),
+        LocationCard("Worcester"),
+        LocationCard("Birmingham"),
+        LocationCard("Birmingham"),
+        LocationCard("Birmingham"),
+        LocationCard("Coventry"),
+        LocationCard("Coventry"),
+        LocationCard("Coventry"),
+        LocationCard("Nuneaton"),
+        LocationCard("Redditch"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Iron Works"),
+        IndustryCard("Coal Mine"),
+        IndustryCard("Coal Mine"),
+        IndustryCard("Coal Mine"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Cotton"),
+        IndustryCard("Pottery"),
+        IndustryCard("Pottery"),
+        IndustryCard("Pottery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+        IndustryCard("Brewery"),
+    ],
+}
+
+STARTING_WILD_LOCATION_CARDS = [
+    # idk how many wilds there are
+    LocationCard("Wild location", True),
+    LocationCard("Wild location", True),
+    LocationCard("Wild location", True),
+    LocationCard("Wild location", True),
+    LocationCard("Wild location", True),
+]
+
+STARTING_WILD_BUILDING_CARDS = [
+    IndustryCard("Wild building", True),
+    IndustryCard("Wild building", True),
+    IndustryCard("Wild building", True),
+    IndustryCard("Wild building", True),
+    IndustryCard("Wild building", True),
+]
