+++ app/environments/brassbirmingham/brassbirmingham/envs/python/id.py
@@ -1,0 +1,5 @@
+import uuid
+
+
+def id():
+    return uuid.uuid1()
