+++ app/environments/brassbirmingham/brassbirmingham/envs/python/print_colors.py
@@ -1,0 +1,30 @@
+def prRed(skk):
+    return "\033[91m {}\033[00m".format(skk)
+
+
+def prGreen(skk):
+    return "\033[92m {}\033[00m".format(skk)
+
+
+def prYellow(skk):
+    return "\033[93m {}\033[00m".format(skk)
+
+
+def prLightPurple(skk):
+    return "\033[94m {}\033[00m".format(skk)
+
+
+def prPurple(skk):
+    return "\033[95m {}\033[00m".format(skk)
+
+
+def prCyan(skk):
+    return "\033[96m {}\033[00m".format(skk)
+
+
+def prLightGray(skk):
+    return "\033[97m {}\033[00m".format(skk)
+
+
+def prBlack(skk):
+    return "\033[98m {}\033[00m".format(skk)
